+++ Kamek/src/world_maps/data/culling_actor.h
@@ -7,7 +7,7 @@
 #define NO_CULL 0.0f
 
 static const f32 wmActorCullRadius[] = {
-    0.0f,     // WM_IBARA        : World 5 vines
+    0.0f,     // WM_IBARA        : World 5 vines (redefined as 0.0f), these use their own culling system
     NO_CULL,  // WM_ANTLION_MNG  : World 2 ambush manager
     250.0f,   // WM_CASTLE       : Castle
     100.0f,   // WM_TOWER        : Unused actor
@@ -25,7 +25,7 @@ static const f32 wmActorCullRadius[] = {
     10.0f,    // WM_STOP         : Unused actor
     170.0f,   // WM_TORIDE       : Tower
     180.0f,   // WM_GHOST        : Ghost House
-    80.0f,    // WM_DOKAN        : Obsolete actor(?) Created but never sets up any model
+    80.0f,    // WM_DOKAN        : Obsolete, likely rendered warp pipe models, however those are baked into the map
     120.0f,   // WM_CANNON       : Warp Cannon
     200.0f,   // WM_KOOPASHIP    : Airship
     100.0f,   // WM_BOSS_BASE    : Base actor for bosses (not used directly)
@@ -51,12 +51,12 @@ static const f32 wmActorCullRadius[] = {
     NO_CULL,  // WM_GRID         : Unused, dummied-out actor
     NO_CULL,  // WM_ISLAND       : World Select island, cannot be culled
     NO_CULL,  // WM_ITEM         : Items menu powerups, cannot be culled
-    2000.0f,  // WM_CS_W3_PALM   : Palm tree manager for World 2 and World 4
+    2000.0f,  // WM_CS_W3_PALM   : Palm tree manager for Worlds 2 and 4
     100.0f,   // WM_SURRENDER    : Tower/Castle clear flags
     800.0f,   // WM_KOOPA_CASTLE : Bowser's Castle
     100.0f,   // WM_ANCHOR       : Airship anchor
     250.0f,   // WM_PAKKUNHEAD   : World 7 climbable vines
-    2500.0f,  // WM_CLOUD        : World 7 clouds
+    2500.0f,  // WM_CLOUD        : World 7 clouds (also use their own culling system)
     400.0f,   // WM_SMALLCLOUD   : World 7 clouds, and the cloud in World 6
     250.0f,   // WM_KOOPAJR      : Bowser Jr
     10000.0f, // WM_PEACH        : Princess Peach